refactor: keep plugin blob fetch local
