+++ nexus-common/src/db/mod.rs
@@ -15,34 +15,3 @@ pub use graph::queries;
 pub use graph::setup;
 pub use graph::GraphOps;
 pub use kv::RedisOps;
-
-/// Fetch the raw bytes for a `pubky://` URI via the pubky SDK.
-///
-/// Shared by the core social watcher and domain plugin handlers so both use
-/// an identical fetch code path.  Returns an error (never silently drops) so
-/// callers can propagate failures to the retry queue.
-pub async fn fetch_blob(uri: &str) -> Result<Vec<u8>, crate::models::event::EventProcessorError> {
-    let pubky = PubkyConnector::get()?;
-    let response = pubky
-        .public_storage()
-        .get(uri)
-        .await
-        .map_err(|e| crate::models::event::EventProcessorError::client_error(e.to_string()))?;
-
-    if !response.status().is_success() {
-        let status = response.status();
-        let body = response
-            .text()
-            .await
-            .unwrap_or_else(|_| "<unable to read body>".to_string());
-        return Err(crate::models::event::EventProcessorError::client_error(
-            format!("Fetch resource failed {uri}: HTTP {status} - {body}"),
-        ));
-    }
-
-    response
-        .bytes()
-        .await
-        .map(|b| b.to_vec())
-        .map_err(|e| crate::models::event::EventProcessorError::client_error(e.to_string()))
-}
